added policy to roles and permissions

+++ app/Http/Controllers/Admin/PermissionController.php
@@ -13,7 +13,8 @@ class PermissionController extends Controller
      */
     public function index()
     {
-        
+        $this->authorize('viewAny', Permission::class);
+
         $permissions = Permission::all();
         return view('permissions.index', compact('permissions'));
     }
@@ -23,6 +24,7 @@ class PermissionController extends Controller
      */
     public function create()
     {
+        $this->authorize('create', Permission::class);
         $permission = Permission::all();
         return view('permissions.create', ['permission' => $permission]);
     }
@@ -32,13 +34,14 @@ class PermissionController extends Controller
      */
     public function store(Request $request)
     {
+        $this->authorize('create', Permission::class);
+
         $data = request()->validate([
             'name' => 'required|string|unique:roles|max:50',
             'description' => 'nullable|string|max:100',
         ]);
     
         Permission::create($data);
-        
         return redirect()->route('permissions.index')->with('success', 'The role has been added.');
     }
 
@@ -47,6 +50,7 @@ class PermissionController extends Controller
      */
     public function destroy(string $id)
     {
+        $this->authorize('delete', Permission::class);
         Permission::find($id)->delete();
         return redirect()->route('permissions.index')->with('success', '204');
     }
